ra8d2_icu_regs + ra_icu: fix NMI + IELSR offsets, 32-bit NMI regs
The ICU header had NMI + IELSR register offsets off by ~0x6000 and
modeled the NMI registers as 8-bit when FSP R_ICU_Type says they
are 32-bit. Verified against FSP R7KA8D2KF_core0.h lines 8670-9061:

Field Was Now Notes
----- --- --- -----
NMICR 0x140 (8b) 0x0010 (8b)
NMIER 0x120 (8b) 0x6100 (32b)
NMICLR 0x130 (8b) 0x6110 (32b)
NMISR 0x114 (8b) 0x6120 (32b)
IELSR[] 0x300 (96x32b) 0x6300 (96x32b)
num_ielsr 112 96

Also fills in the surrounding IRQCRa[16]/NMICR/IRQCRb[16]/INTSELR[32]
offsets so the header documents the full front end of the ICU block
(FSP splits IRQCR into an `a` copy at 0x00 for core0 and a `b` copy
at 0x14 for core1; we rename the symbol to `irqcra0` to match).

API change in ra_icu.h: ra_icu_nmi_enable / disable / clear take
uint32_t mask (was uint8_t) and ra_icu_nmi_status writes uint32_t*
(was uint8_t*). The FSP NMIER bits go 0..20 (IWDTEN, WDTEN, PVD1/2,
SOSTEN, OSTEN, NMIEN, BUSEN, CMEN, LMEN, LUEN, FPUFLTEN, MRCRDEN,
MRERDEN, IPCEN) so 8-bit was insufficient for the real register.

Every peripheral interrupt routes through IELSR, so until this
landed no NVIC line could be configured -- the old 0x300 base hit
a reserved area. Combined with the PFS PWPR fix in the previous
commit, the two cross-cutting interrupt-control blockers are now
unblocked.

Tests: test_ra_icu.c updated to:
- Use 96-slot IELSR range (last valid = 95, bad = 96).
- Test NMI with 32-bit mask values covering bits 0..20.
- Pre-pollute NMIER with 0x1FFFFF instead of 0xFF.
65/65 host tests pass; cross build clean; cite_check + world tags
both green.

+++ libs/ra_hal/inc/ra_icu.h
@@ -172,19 +172,19 @@ typedef struct {
  *       register via ``ra_icu_nmi_disable``.
  * @since 0.2.0
  */
-[[nodiscard]] ra_err_t ra_icu_nmi_enable(uint8_t mask);
+[[nodiscard]] ra_err_t ra_icu_nmi_enable(uint32_t mask);
 
 /**
  * @brief Disable NMI sources by clearing bits in NMIER.
  *
- * @param[in] mask Bit mask to clear in NMIER.
+ * @param[in] mask Bit mask to clear in NMIER (21 bits, FSP R_ICU_NMIER).
  *
  * @return ``k_ra_ok``.
  * @pre IRQs masked or single-threaded init context.
  * @post NMIER has the requested bits clear.
  * @since 0.2.0
  */
-[[nodiscard]] ra_err_t ra_icu_nmi_disable(uint8_t mask);
+[[nodiscard]] ra_err_t ra_icu_nmi_disable(uint32_t mask);
 
 /**
  * @brief Clear pending NMI status flags via NMICLR (HUM 14.2.15 p 544).
@@ -196,19 +196,19 @@ typedef struct {
  * @post The targeted NMISR bits read as 0.
  * @since 0.2.0
  */
-[[nodiscard]] ra_err_t ra_icu_nmi_clear(uint8_t mask);
+[[nodiscard]] ra_err_t ra_icu_nmi_clear(uint32_t mask);
 
 /**
  * @brief Read the current NMISR value.
  *
- * @param[out] out_status NMISR value on success.
+ * @param[out] out_status NMISR value on success (32-bit).
  *
  * @return ``k_ra_ok`` or ``k_ra_err_null_ptr``.
  * @pre ``out_status`` is non-NULL.
  * @post No hardware state is modified.
  * @since 0.2.0
  */
-[[nodiscard]] ra_err_t ra_icu_nmi_status(uint8_t* out_status);
+[[nodiscard]] ra_err_t ra_icu_nmi_status(uint32_t* out_status);
 
 /* =============================================================================
  * Legacy API -- direct NVIC + IELSR access (deprecated)
